Fixed typos and separated logic for concrete buildings
Issue #31
